+++ src/components/page/homeHeader.tsx
@@ -1,19 +1,18 @@
 import React, { Suspense, lazy, useEffect } from 'react'
-import Switch from 'react-switch'
 import { connect } from 'react-redux'
 import { factionNames, selections, selectors, realmscape } from 'ducks'
 import { useTheme } from 'context/useTheme'
 import { useSavedArmies } from 'context/useSavedArmies'
 import { useAppStatus } from 'context/useAppStatus'
-import { centerContentClass } from 'theme/helperClasses'
 import { withSelectOne } from 'utils/withSelect'
 import { logFactionSwitch } from 'utils/analytics'
 import { componentWithSize } from 'utils/mapSizesToProps'
 import { titleCase } from 'utils/textUtils'
 import { getArmyLink } from 'utils/handleQueryParams'
+import { LinkNewTab } from 'components/helpers/link'
 import { LoadingHeader } from 'components/helpers/suspenseFallbacks'
 import { SelectOne } from 'components/input/select'
-import { LinkNewTab } from 'components/helpers/link'
+import ToggleGameMode from 'components/input/toggle_game_mode'
 import { SUPPORTED_FACTIONS, TSupportedFaction } from 'meta/factions'
 import { IStore } from 'types/store'
 
@@ -115,39 +114,6 @@ const JumbotronComponent: React.FC<IJumbotronProps> = props => {
             </div>
           </>
         )}
-      </div>
-    </div>
-  )
-}
-
-const ToggleGameMode = () => {
-  const { theme } = useTheme()
-  const { isGameMode, toggleGameMode } = useAppStatus()
-
-  const spanClass = `align-self-center pb-2`
-
-  return (
-    <div className={`${centerContentClass} ${theme.text}`}>
-      <div className={`d-inline-flex flex-row`}>
-        <span className={`${spanClass} mr-2 ${isGameMode ? `` : `font-weight-bold`}`}>Edit</span>
-        <label htmlFor="visual-theme-switch" className="mb-0">
-          <Switch
-            onChange={toggleGameMode}
-            checked={isGameMode}
-            onColor="#1C7595"
-            onHandleColor="#E9ECEF"
-            handleDiameter={30}
-            uncheckedIcon={false}
-            checkedIcon={false}
-            boxShadow="0px 1px 5px rgba(0, 0, 0, 0.6)"
-            activeBoxShadow="0px 0px 1px 10px rgba(0, 0, 0, 0.2)"
-            height={20}
-            width={80}
-            className="react-switch"
-            id="visual-theme-switch"
-          />
-        </label>
-        <span className={`${spanClass} ml-2 ${isGameMode ? `font-weight-bold` : ``}`}>Play</span>
       </div>
     </div>
   )
